pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "stac_fastapi.globus_search"
version = "0.0.1"
description = "A Globus Search stac-fastapi frontend"
license = "Apache-2.0"
authors = ["Stephen Rosen <sirosen@globus.org>"]
packages = [
  { include = "stac_fastapi", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12.0"
attrs = "25.3.0"
globus-sdk = "3.62.0"
hishel = {extras = ["async"], version = "1.1.5"}
pystac-client = "0.7.0"
sfeos-helpers = "6.3.0"
"stac-fastapi.api" = "6.0.0"
"stac-fastapi.extensions" = "6.0.0"
"stac-fastapi.types" = "6.0.0"
stac-fastapi-core = "6.3.0"
uvicorn = "0.35.0"

[tool.poetry.scripts]
stac-fastapi-globus-search = "stac_fastapi.globus_search.app:run"
